requirements.txt
streamlit
groq
pydantic
python-dotenv
requests
